[RENAME] components/cards로 이동 및 이름 수정
